Renamed logging_utils module
