+++ packages/opal-common/opal_common/logger.py
@@ -4,10 +4,10 @@ import sys
 from loguru import logger
 
 from opal_common.config import opal_common_config
-from opal_common.logging.filter import ModuleFilter
-from opal_common.logging.formatter import Formatter
-from opal_common.logging.intercept import InterceptHandler
-from opal_common.logging.thirdparty import hijack_uvicorn_logs
+from opal_common.logging_utils.filter import ModuleFilter
+from opal_common.logging_utils.formatter import Formatter
+from opal_common.logging_utils.intercept import InterceptHandler
+from opal_common.logging_utils.thirdparty import hijack_uvicorn_logs
 from opal_common.monitoring.apm import fix_ddtrace_logging
